requirements.txt
faster-whisper
PyQt6
google-generativeai
python-dotenv
ffmpeg-python
deep-translator
